Initial asset importer
