+++ UnitySDK/Assets/ML-Agents/Scripts/Demonstration.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace MLAgents
+{
+    public class Demonstration : ScriptableObject
+    {
+        [SerializeField] public string demonstrationName;
+    }
+}
