+++ UnitySDK/Assets/ML-Agents/Scripts/Demonstration.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b651f66c75a1646c6ab48de06d0e13ef
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
