+++ UnitySDK/Assets/ML-Agents/Scripts/DemonstrationRecorder.cs.meta
